The job size was not long long in the thread struct
Because of this the number of chunks shot up every time the size was higher than 10gigs

+++ utils.h
@@ -6,7 +6,7 @@ using namespace std;
 struct Job {
 	int thread_id ;
 	vector<string> block_names;
-	int block_size;
+	unsigned long long block_size;
 	int chunk_size;
 	time_t elapsed_time ;
 };
